fix: restore npm SIGINT behavior (reverts #41) (#52)
This reverts the behavioral change from #41.

This change caused pnpm to improperly handle interactive
shell Ctrl-C, trapping the received signal and also ignoring
the abnormal exit of the script. This often results still
in a quick exit (because the script has terminated), but
with exit code 0.

PR #41 resulted in incorrect behavior, which also diverges
from npm (as well as `node --run`). As another problem, it
made pnpm internally inconsistent as Ctrl-C results in non-zero
in situations when lifecycles scripts are not actively running.
Even in the context of pnpm run, a very quick Ctrl-C, prior to
spawn, results in a non-zero pnpm exit.

For additional discussion see: #51

Fixes: pnpm/pnpm#9626